feat: add posts section
Add a <section> tag with the class and id posts. Inside it, add a <div> tag with the classes post and post-1. Inside the <div>, add an <a> tag. Inside the <a> tag, add a <div> tag with the class title. Inside this <div>, add an <h3> tag and a <p> tag